swizzle the ApiItem and add our custom DocItem swizzles to it, add a line back to docusaurus.config, and delete DocItem
